fix(P2BI-201): rename files
