+++ src/index.ts
@@ -2,7 +2,7 @@ import express from 'express';
 
 import healthRoute from './routes/health.route';
 import etBlogRoute from './routes/etBlog.route';
-import etnewsRoute from './routes/etnews.route';
+import etNewsRoute from './routes/etNews.route';
 
 const app = express();
 app.use(express.json());
@@ -11,7 +11,7 @@ const PORT = process.env.PORT || 8080;
 
 app.use('/health', healthRoute);
 app.use('/et-blog', etBlogRoute);
-app.use('/et-news', etnewsRoute);
+app.use('/et-news', etNewsRoute);
 
 app.listen(PORT, () => {
     console.log(`Server is running on http://localhost:${PORT}`);
