+++ src/services/etNews.service.ts
@@ -1,5 +1,5 @@
 import db from '../utils/db.util';
-import { ETNews } from '../types/etnews';
+import { ETNews } from '../types/etNews';
 
 
 export default {
