+++ src/types/etNews.ts
